arm: new labels in the linker script file
_bss_start_ofs is used in start.S to indicate end of copied
image. This may not be correct when we have a discontiguous
memory map. For instance, .bss may be placed in SDRAM for
some SPLS while rest of the image is placed in SRAM.

Define a new label in linker script to indicate the end of the
image copied during relocation and use it appropriately in
start.S.

Also, add a new label to indicate the end of flash image.
This will be useful in identifying the size of flash image

Signed-off-by: Aneesh V <aneesh@ti.com>