omap: spl: add more debug traces
In SPL console is enabled very early where as in U-Boot
it's not. So, SPL can have traces in early init code
while U-Boot can not have it in the same shared code.

Adding a debug print macro that will be defined in SPL
but compiled out in U-Boot.

Also add some useful debug traces throughout SPL

Signed-off-by: Aneesh V <aneesh@ti.com>