OMAP4: charger: Cancel charging if chip is not detected
Sometimes device hangs on boot if an error occurs
on reading i2c bus in twl603x battery charger driver.
At that case the chip type TWL6030 or TWL6032
is not determined and the driver assumes the default
value. It's not the safe action.
The i2c errors could corrupt following bus read/write
operations and cause an unstable behaviour of twl603x.
In this patch added cancelling the further interaction
with twl603x in this driver if such a fault on i2c bus
is detected. Thus if i2c error occurs we'll cancel
the calibration, GPADC measuring of battery presence
and voltage and the charger start/stop.
When the kernel boots it uses another more complex
driver which is not impacted by the errors like this
and performs the same functions.

Change-Id: Idb3c1b7ff1a6e99d0a5b785ba5e3507399ca1a10
Signed-off-by: Oleg_Kosheliev <oleg.kosheliev@ti.com>