arm: omap4+: pmic: Make generic bus init and write functions
Voltage scaling can be done in two ways:
-> Using SR I2C
-> Using GP I2C
In order to support both, have a function pointer in pmic_data
so that we can call as per our requirement.

And also fixing a bug which is introduced due to recent cleanup.
On 4460 TPS supplies only to vdd_mpu. Twl6030 supplies voltage to
vdd_core, vdd_mm. Correcting the same.

Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>