ARM: dra: emif: Add different EMIF enable configs
There are some boards where EMIF1 and 2 together
are not working, but if only EMIF1 or 2 is enabled,
it improves the situation. So let us add configs
to have both EMIF1 and 2, EMIF1 or EMIF2 only.

The default will be both EMIF1 and EMIF2. So
if there is a instability, first is to change this
and test.

Signed-off-by: Sricharan R <r.sricharan@ti.com>