Add resource table processing with associated L2 MMU programming.
The ELF executable is searched for a .resource_table section.  If found,
program the L2 MMU and handle the different resources.
Use page table instead of TLB and allocate from CMA reserved area

The TLB isn't big enough to hold all the translations for DEVMEM
entries, so move to a 1-level page table with 1 MB "sections".
The page table (16 KB) is placed at the end of the CMA reserved area.
It's possible that this location is needed by the firmware (in which
case the firmware is using pretty much *all* of the reserved area), but
there doesn't seem to be a better location to place it.

Replicated the CMA allocation scheme that allocates memory from the CMA
area that is reserved for IPU1 in the Linux kernel.

Change-Id: I80b04e7b232e7f93d24984c44064bca3599a9223
Signed-off-by: Robert Tivy <rtivy@ti.com>