Enable loading of IPU image from mmc card
Add the functionality of loading IPU image from first stage of the
boot loader. The image is loaded from the fat partition of the mmc
card. This adds the elf parser to load the IPU image.
The code is defined under the config CONFIG_LATE_ATTACH_BOOTIPU1
and this config is mutually exclusive with the CONFIG_BOOTIPU1.

Change-Id: I418c44dbab8d05f6c42bd8f8ff16f928a96efd0c
Signed-off-by: Amarinder Bindra <a-bindra@ti.com>