dra7xx: Add support for loading IPU2 from the boot loader
This adds support for loading the IPU2 elf binary from u-boot and sets up the
iommu, cma allocation for IPU2. This is based on the IPU1 loading functionality
for late attach by the remoteproc driver in the kernel.
The code for loading IPU2 is defined under config CONFIG_LATE_ATTACH_BOOTIPU2.
By default the functionality is disabled, enable the config to use the loading
functionality for IPU2 from u-boot.

Change-Id: I38125c054e8716585c4af64f089cfaff0263f40c
Signed-off-by: Amarinder Bindra <a-bindra@ti.com>