Add header file to enable early boot of Ipc-enabled executable
Add the file remoteproc.h which is used to enable the boot of
the remote core like IPU from the boot loader. This file includes
the declarations and macros which would be required to enable the
loading and programming of the remote core based on Ipc resource
table entry from the elf file.

Change-Id: Id845c71fd309f5254604849505b7932b6a7733a8
Signed-off-by: Robert Tivy <rtivy@ti.com>
Signed-off-by: Amarinder Bindra <a-bindra@ti.com>