spl: DRA7XX: enable late attach feature for eMMC SPL boot
The late attach for both the IPU1 and IPU2 cores is enabled
for the eMMC single stage bootloader use case. This will enable
the single stage boot loader to load the IPU binary if it exists
in the eMMC partition.

Currently, only one of the cores, IPU1 or IPU2 can be loaded using the
single stage bootloader.

While at it, fixed minor cosmetic change.

Change-Id: I648a697b195b996802bd671df834176c9dd1315e
Signed-off-by: Subramaniam Chanderashekarapuram <subramaniam.ca@ti.com>