dra7xx: late attach: Update the IPU resource table
Once the boot loader has done the iommu configuration and
handled the resources in the resource table, some information
like the physical addresses for the carveouts, etc needs to be
feeded back to the remote core as it may require this information.
Therefore, added a copy of the updated table by the boot loader
to the actual resource table in the memory.

Change-Id: I66432fe9bf1a370110d95d39993ddeddbbd61f2e
Signed-off-by: Robert Tivy <rtivy@ti.com>
Signed-off-by: Amarinder Bindra <a-bindra@ti.com>