ARM: DRA7: Add 667MHz only for DRA72x
With the current settings, DRA75x is also defaulted to 667MHz,
but it should be applicable only for DRA722. Fixing the same.

Signed-off-by: Lokesh Vutla <lokeshvutla@ti.com>

Conflicts:

    arch/arm/cpu/armv7/omap5/hw_data.c
    arch/arm/cpu/armv7/omap5/sdram.c

[port to p-ti-u-boot-2013.04 and add change to android config]
Change to use only for DRA722 was already present with the
previous patch, so only took the part of patch which removed
the CONFIG_DDR_666MHZ flag.

Change-Id: If04ae531f53c83b5da08626d54d63d808ca86401
Signed-off-by: Angela Stegmaier <a0866189@ti.com>