Set vring 'da' (device address) to allocated 'pa' (physical address)
The vrings in the loaded copy of the resource table need to be updated
with the physical address of the vring's allocated memory.  This
overwrites the virtual address with which the vring resource table entry
is initialized when loaded.

This matches equivalent behavior in the Linux remoteproc (for both late
and normal attach modes). It also allows the expected value from
Resource_getVringDA() in Ipc 3.22.00.05 and later to be correct.

This has thus far not generally been observed as a problem in the early
boot case since the vrings can't be actively used until the Linux
remoteproc comes alive, and since remoteproc initializes the vrings
for both late and normal attach modes, the error in u-boot gets fixed
up by the time the vrings are used on the remote core.

Change-Id: I3966300a80a0f1a740a832f32b71962f5def8633
Signed-off-by: Robert Tivy <rtivy@ti.com>