fastboot: OUT transaction length must be aligned to wMaxPacketSize
OUT transactions must be aligned to wMaxPacketSize for each transfer,
or else transfer will not complete successfully. This patch modifies
rx_bytes_expected to return a transfer length that is aligned to
wMaxPacketSize.

Note that the value of ep->desc->wMaxPacketSize and ep->maxpacket
may not be the same value, and it is the value of ep->desc->wMaxPacketSize
that should be used for alignment.

Change-Id: I3f9b2f12502bb0ffa96fc4244e50cb74b62ac74d
Signed-off-by: Shankar Rao <shankar.nrao@ti.com>
Signed-off-by: Angela Stegmaier <angelabaker@ti.com>