dra7xx_evm: Enable IP_DEFRAG to reduce tftp time.
This change allows the increase of tftp blocksize from 1468
to 16268. This increases the tftp data transfer rate for zImage
from 1.1 MB/s to 6.1 MB/s.

The downside of this change is that ACK's will now be sent once for the
whole packet size(16KB) instead of the earlier 1468 bytes. Loss of any
fragment of the packet requires that the whole 16 KB be retransmitted.
However on a local network, the probability of packet loss is very low
and turning this flag on is justified.

Change-Id: I213df76259fe0a18856984cfbeab95c627d7a6ee
Signed-off-by: Venkateswara Rao Mandela <venkat.mandela@ti.com>