usb: gadget: fastboot: Add fastboot erase
Add the functionality for fastboot erase. Fastboot erase
will erase a partition which is specified by name. The
erase is performed based on erase group size, so to avoid
erasing other partitions, align the start address and size
to the erase group size for the device.

Currently only supports erasing from eMMC.

Change-Id: I9a514a0466f38740fd47c40f16771cdf5de90340
Signed-off-by: Angela Stegmaier <angelabaker@ti.com>