fastboot: Remove references to do_mmcops
Cleanup the fastboot files to remove references to do_mmcops
and instead directly call the mmc apis, as is done in other
parts of the code. With this change, cmd_mmc no longer
needs to be included when including fastboot in the build.

This is useful when building fastboot into the SPL,
as it reduces the size of the SPL.

Change-Id: I6046d8699df07cb8ade578191ba5723d071a3fc4
Signed-off-by: Angela Stegmaier <angelabaker@ti.com>
Signed-off-by: Ravi Babu <ravibabu@ti.com>