u-boot: configs: dra7xx_evm_android: Move Download Address
The ramdisk size for LP can be larger than the 1MB that is
currently available when the download address is 0x82000000.
The download address is used to download the header from
the boot.img (or the recovery.img). The ramdisk load address
is set to 0x81f00000, and the size of the ramdisk in the
case of recovery is larger than 1MB, meaning that after
downloading, it will overwrite the header. This causes the
rest of the boot process to use incorrect header information.

To fix this, add another 1MB to the download address, to
prevent a conflict with the ramdisk.

Change-Id: I2b5a6cd00f742950151131a9941c377da02bed9f
Signed-off-by: Angela Stegmaier <angelabaker@ti.com>