omap5/dra7xx: implement get_sysboot_value()
The boot device settings supplied by boot rom to SPL depend on the boot
device order and the boot media in which a valid SPL binary was found.
There might be cases during development where we would like to use the
actual boot device settings on the sysboot pins e.g. determining whether
we are in production boot mode or development boot mode. This function
returns the value of sysboot pins [5:0] for DRA7XX and OMAP5.

Change-Id: Ia4172419471930980baaaffbc152f17e5e6150f1
Signed-off-by: Venkateswara Rao Mandela <venkat.mandela@ti.com>