dra7xx: spl_fat: look for dra7-evm.dtb when loading OS
Currently SPL looks for the file "args" before loading the OS. This
appears to be needed for passing boot arguments. As we are using the
device tree now and can pass boot arguments via the "chosen" node, I
have modified the code to look for the file dra7-evm.dtb.

Change-Id: I55fe59854b0c1566f2b569eb4e6c56ce9bb63ed1
Signed-off-by: Venkateswara Rao Mandela <venkat.mandela@ti.com>