dra7xx: spl: set default return value of spl_start_uboot to 1.
spl_start_uboot() is expected to return 0 for OS loading and a non-zero
value for starting u-boot.

If the default return of spl_start_uboot() is 0, SPL tries to load the
OS, fails and then falls back to U-Boot. Setting the default return
value to 1 allows spl_start_uboot() to only capture the conditions where
the OS needs to be booted directly.

Change-Id: Ib27b9cb8323af5c440513e48f6802e107e85e6f3
Signed-off-by: Venkateswara Rao Mandela <venkat.mandela@ti.com>