configs: dra7xx_evm_android: add ipu2 partition
Rename the ipu partition to ipu1 and add an ipu2 partition.
This is so that seperate images can be loaded to IPU1 and IPU2
for late attaching.

If the late attach flag is not defined or if there is no image
loaded in the partition, then no early loading with happen. In
this way, the late attach flag can be defined, but if there is
only an image loaded in the ipu1 partition, and not in the ipu2
partition, then only the IPU1 image will be loaded from the
bootloader.

NOTE: This change requires a change to ramdisk in Android file system.

Change-Id: I306e2e7dcaae3aec689138e8403a13d6b1729d63
Signed-off-by: Angela Stegmaier <angelabaker@ti.com>