spl: dra7xx: Passing the efused MAC address via DT in single stage boot
Read the MAC address from EFUSE and pass it to the Linux kernel via DT
in single stage boot. Without this change, the kernel will receive an
all zero mac address from DT. This will result in a random mac address
being used on each reboot and may lead to exhaustion of the DHCP IP pool
over multiple reboots of the target.

Change-Id: I18dfd56fd78ec736d211919f03c74e9738964c19
Signed-off-by: Venkateswara Rao Mandela <venkat.mandela@ti.com>