u-boot: security: Add image verify infrastructure
This patch adds the code to support image authentication
on secure devices using the ROM HAL APIs accessed via
the SMC assembly instruction. Code that relies on this
infrastructure will be added in later patches.

The added secure_boot_verify_image API can be utilized
for authenticating any image which was signed using
the ift-image-sign.sh script that is delivered as part
of the M-Shield DK (v4.3.0 or later).

Change-Id: I74b38fee278912b17cee63c967c4c6de2251f071
Signed-off-by: Daniel Allred <d-allred@ti.com>