u-boot: security: Add image verifcation for u-boot
This patch adds support for the SPL to authenticate
a signed u-boot.img. This authentication is enforced,
meaning that boot will fail if they authentication
does not pass. This authentication step is the first
link in the chain of authentication that end-user may
choose to enforce. The chained authenication is enabled
by defining CONFIG_SECURE_BOOT.

1. Secure configs updated to force CONFIG_SECURE_BOOT
   to be defined.
2. Makefile updated so that u-boot.img is processed by
   the ift-sign-image.sh script from the M-Shield DK
   when CONFIG_SECURE_BOOT and CONFIG_OMAP_SECURE are
   both defined. The resulting u-boot.img has a valid
   digital signature appended to the end of the binary.
3. Add code to spl.c main loop to enforce the signature
   check using the secure_boot_verify_image API. Boot
   will halt if the image is not verified.

Change-Id: Ifd1b5efa16cb8236d9ffc5343a676765519e6aaf
Signed-off-by: Daniel Allred <d-allred@ti.com>