u-boot: security: Fixes for omap secure services
With DCache enabled (as it is by default in the main
u-boot image, but not in the SPL), calls from the
non-secure world to the secure world (via the SMC
instruction) would fail since the content and parameters
that were being passed to the secure world were stuck
in the cache and the non-secure cache lines are not visible
to the secure world. This patch updates the parameter passing
to use a fixed buffer which is flushed before the SMC call.
The verify API is also updated to flush the content buffer
to be verified.

Change-Id: I9d74428ede5f2c85d756d2c01b719a9ba8c16d33
Signed-off-by: Daniel Allred <d-allred@ti.com>