spl: early boot: use same page table for both DSP MMU's
On DRA7xx, each DSP has two MMU's. one for the CPU access path and one
for the EDMA access path. Earlier u-boot code maintained separate but
identical page tables for both MMU's. This is unnecessary.

- The remoteproc kernel module, with which the u-boot code should
  interoperate, does not support distinct page tables for both
  MMU's.

This commit modifies the data structures and code in u-boot to use the
page table address for both the DSP MMU's. Using a single page table for
both MMU's reduces the amount of memory used for the page table to 16 KB
from 32 KB.

Change-Id: Ie8ec855c6156b25b98b988668cb88cd292c8471f
Signed-off-by: Venkateswara Rao Mandela <venkat.mandela@ti.com>