ARM: DRA7: Add Workaround for DRA7 DSP MFLAG Errata i872
This patch adds the workaround for Errata i872, which affects
the DSPs on DRA75x/74x ES1.0 and ES1.1.

Errata Description:

"i872: DSP MFlag Output Not Initialized

The DSP1 and DSP2 Subsystems include MFlag output signals that
are under DSP software control and are used to control arbitration
at various points in the system interconnect, including in the
DMM and EMIF command queues.
Each DSP subsystem's MFlag output is uninitialized in hardware
until the DSP is powered up and clocked, and can default to a value
of either 0 or 1. This can have unanticipated and non-deterministic
effects on system traffic dependent on the power-on state of the
MFlag signals"

Workaround:

To ensure a known value for the DSP(s)'s MFlag output, the software
should power up and clock the DSP(s) for a brief time. After doing
this, the MFlag output will be 0. The DSP(s) may be immediately
disabled afterward.

This patch adds code to the board init sequence for DRA7X to enable the DSP1
(and DSP2 if supported) clocks and bring the DSP out of reset.
After validating that the DSP is enabled, the clocks are again disabled
and the DSP is put back to reset. The sequence is taken from the errata
documentation, except that the DSP is not put into power OFF state after
performing the workaround. Since the initial state of the DSP at reset
prior to executing the errata workaround is power ON, the sequence
leaves the DSP in the same power state. It is up to subsequent software
to put the DSP to power OFF state, if desired. In this way the DSP is
left in the same state as before the workaround is run.

Please see DRA75x/DRA74x errata documentation (SPRZ398F) section 1.48,
"DSP MFlag Output Not Initialized," for complete details.

Change-Id: Id0d6f8bc76a1de86199656bcc29b9b0cf15bc980
Signed-off-by: Angela Stegmaier <angelabaker@ti.com>