ARM: DRA74-evm: Add iodelay values for SR2.0
Silicon revision 2.0 has new signal routing hence has an updated set of
iodelay parameters to be used. Update the configuration for the same.
Padmux remains the same.

Based on data from VayuES2_EVM_Base_Config-20150807.

NOTE: With respect to the RGMII values, the Manual IODelay values
are used for the fine adjusments needed to meet the tight RGMII
specification.

Change-Id: I37a57e86b51685a9c02df98ed032a5ae49ace250
Signed-off-by: Nishanth Menon <nm@ti.com>
[Back ported to 2014 u-boot]
Signed-off-by: Vishal Mahaveer <vishalm@ti.com>