spl: early boot: fix timing issue with DSP power on.
Early boot code assumed that the DSP1/2 RSTST registers contain their
cold reset values when the early boot code is executing.

Since the commit

ARM: DRA7: Add Workaround for DRA7 DSP MFLAG Errata i872
add0afa84900e89595bfe1e6999b42b719031bcb

the assumption is not true. Some of the bits in RM_DSP1/2_RSTST register
were set as a result of powering on the DSP, taking the DSP out of reset
for a short time, putting it back to reset but not powering it off.
This resulted in status checks during DSP1/2 early boot passing
prematurely and caused instability.

This commit clears the status bits before bringing the DSP MMU,Cache and
slave interface out of reset and fixes the issue.

Change-Id: Icffc50a82262c9292abe414270c9ab958478d320
Signed-off-by: Venkateswara Rao Mandela <venkat.mandela@ti.com>