ti: common: board_detect: Add ability to setup board information
Kernel stores information to the RTC_SCRATCH0 and RTC_SCRATCH1
registers for wakeup from RTC-only mode. Parse these registers during
SPL boot and jump to the kernel resume vector if the device is waking
up from RTC-only mode

During the normal boot patch the SCRATCH1 : bits16-31 are updated
with the eeprom read board type data. In the rtc_only boot path the
rtc scratchpad register is read and the board type is determined and
correspondingly ddr dpll parameters are set. This is done so as to
avoid costly i2c read to eeprom.

So, we update the eeprom data store region with data corresponding to
this.

Change-Id: I83bce154420f16a7dc83dce97f2520dd7fb1f61e
Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Ravi Babu <ravibabu@ti.com>