u-boot: security: New SPL entrypoint for HS parts
This patch updates the entry point address to be 4K beyond
the current value. This change is required for compatibility
with the latest TI Secure Development Package for the DRA7xx
devices. That release reserves the first 4KB of the
OCMC RAM for secure world usage, meaning the public world
u-boot SPL code cannot be loaded there. This change only
impacts builds for secure devices.

Change-Id: I86ea6d75b8a2e1825c38e2001ca0cce4224217c7
Signed-off-by: Daniel Allred <d-allred@ti.com>