u-boot: security: Add APIs for DDR firewalls setup
Adds support for new EMIF firewall APIs that are intended
to be used for static (one-time) config of firewall regions
in the EMIF memory space. These APIs are new in the latest
TI Secure Development Package and exist to allow static config
of the firewall before the EMIF space is in use. As such, they
should be used as part of the EMIF setup that happens in the
u-boot SPL code.

Change-Id: I32e57c5c6c9770062cb62365819e755121cfd2a2
Signed-off-by: Daniel Allred <d-allred@ti.com>