u-boot: security: Update fdt on HS devices
Added code to update the binary FDT in memory so that
changes made to the secure memory reservation in the u-boot
config can properly pass up to the booting kernel. This
means that there is no need for HS specific device tree files
or overlays needed in the kernel tree.

Disabled the DTB authentication since the dynamic updates described
above break it. To fix that in the long term, some reordering of
steps will have to happen (authenticate before in-memory mods).

Change-Id: I1d9c9e21d66fa8e8472b3873fcb141e8c3ca1d2e
Signed-off-by: Daniel Allred <d-allred@ti.com>