arm: dra7: Add iodelay support for different MMC modes
Currently, on the dra7-evm and dra72-evm, the MMC instances
are used in the following modes:
MMC1 - HS Pad loopback
MMC2 - DDR50 Pad loopback

But the iodelays being programmed are for the legacy mode.
To guarantee timings with u-boot drivers, update the pad and
delay arrays to correctly choose the pad/delay configuration
at compile time.

Change-Id: Ic47264f53000fde91311a0f3f02cf92ba8dd1cdd
Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>
[fix Conflicts in board/ti/dra7xx/mux_data.h]
Signed-off-by: Praneeth Bajjuri <praneeth@ti.com>