DRA7XX - Configs - Add config option for LATE_ATTACH_DMA_POOL
Add the config option for LATE_ATTACH_DMA_POOL in the dra7xx_evm
and dra7xx_evm_android config files and set the default to
have it be undefined so that the normal CMA pool is used by default.

To use the DMA pool option, define the CONFIG_LATE_ATTACH_DMA_POOL flag
in the config file and also set the reserved mem pool for the late
attached cores to "no-map" in the kernel dtb file.

Change-Id: I7bea1585cc3487600d17eca2cb434c437f419fd0
Signed-off-by: Angela Stegmaier <angelabaker@ti.com>