drivers: mmc: omap: keep HS flag enabled in host caps
Fixes [2920896 dra7xx: mmc: Selectively enable DDR_52 mode]

Looks like the mmc core is checking for MMC_MODE_HS in controller
capabilities even in case of higher speed modes such as HS_52 or
DDR_52 so keep MODE_HS always enabled.

This will not restrict the speed as when the core finds that
the host and the card both supports HS_52 or DDR_52
it will switch the frequency to that of the higher speed mode.

Change-Id: I1ad01f9f3ea7b4050f6fec59b57d235561f6e8b3
Signed-off-by: Ravikumar Kattekola <rk@ti.com>