ti_omap5_common: Respect USB controller number in fastboot
On "fastboot reboot-bootloader" we check "dofastboot" variable and do
"fastboot 0" command in U-Boot if it's 1. But there is boards which have
USB controller number other than 0, so it should be respected when
performing "fastboot" command.

This patch adds CONFIG_FASTBOOT_USB_DEV option which can be used to
provide correct USB controller number to "fastboot" command.

Change-Id: I7c98f7f5d246070864fc594c7b09e6bb1d332a65
Signed-off-by: Sam Protsenko <semen.protsenko@linaro.org>